Move MembersPage to separate file
